Use channel_id instead of channel to get id
Previously, this line caused a crash when sending responses in a guild
that uses the applications.commands scope. In this case,
SlashContext.channel is None and SlashContext.channel_id has the
channel id.

This appears to be a holdover from pre-1.0.9 when SlashContext.channel
represented either a channel or a channel id.