Use guild_id instead of inspecting guild
Previously, this line caused guild-specific commands to not
execute in guilds with the applications.commands scope. In this case,
SlashContext.guild is None and SlashContext.guild_id has the guild id.

This appears to be a holdover from pre-1.0.9 when SlashContext.guild
represented either a guild or a guild id.