docs: Adjust documentation, fix typos, update locale. (#790)
* ADMIN/revert: remove unwanted file.

This file was actually cloned outside of the client directory in a PR that had not appropriately adjusted to the new architectural layout of the library.

* chore: Update requirements.txt

* chore: Version bump.

* docs: Adjust documentation linking, fix typos.

* docs: Fix message intent reference.

* chore/ADMIN: force version onto old pypi

This is an executive decision made between Delta and I. We will be forcing a version 4.0 push onto our old branch to help transition old users using v3 to our new project. This does not necessarily mean this branch will now be receiving updates along with the rest. The goal is to make our current branch used as much as possible.

* chore/ADMIN: fix content type

* fix/ADMIN: correct improper description

* revert/ADMIN: change back to original

* docs: Fix Client reference.

* docs: Remove Modal beta error tag, fix grammar typo

* docs: Include private objects by default.

* docs: Rebuild .mo/.po localisation for locales.

This may wipe current translation, however the lib docs are different now than the last build (~25/9/2021)

* ci: correct from checks.

* docs: Refactor language build default for PR review, remove :inherit: tag.

* docs: Include rest of locale (ext, context, etc).

This was cut off from some random PyCharm setting.

* ci: correct from checks.

* docs: Fix codeblocks in some areas, fix gateway documentation

* docs: Implement barebones better search capability

This seems to, according to docs, only work when the server's live. But on a PR view it acts as if this commit doesn't do anything.

* docs: Rebuild locale.

* ci: correct from checks.

* docs: Correct models.command.po locale.

* docs: Adjust note/warning label spacing on component.py

* docs: Adjust return doc spacing, rebuild locale.

Co-authored-by: fl0w <41456914+goverfl0w@users.noreply.github.com>
Co-authored-by: Sofia <41456914+ffl0w@users.noreply.github.com>
Co-authored-by: James Walston <41456914+jameswalston@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>