refactor: make `get_role` more consistent  (#865)
* ADMIN/revert: remove unwanted file.

This file was actually cloned outside of the client directory in a PR that had not appropriately adjusted to the new architectural layout of the library.

* chore: Update requirements.txt

* chore: Version bump.

* chore/ADMIN: force version onto old pypi

This is an executive decision made between Delta and I. We will be forcing a version 4.0 push onto our old branch to help transition old users using v3 to our new project. This does not necessarily mean this branch will now be receiving updates along with the rest. The goal is to make our current branch used as much as possible.

* chore/ADMIN: fix content type

* fix/ADMIN: correct improper description

* revert/ADMIN: change back to original

* Update guild.py

* ci: correct from checks.

Co-authored-by: fl0w <41456914+goverfl0w@users.noreply.github.com>
Co-authored-by: DeltaX <33706469+DeltaXWizard@users.noreply.github.com>
Co-authored-by: Sofia <41456914+ffl0w@users.noreply.github.com>
Co-authored-by: James Walston <41456914+jameswalston@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>